+++ releasenotes.md
@@ -1,5 +1,10 @@
 # Release Notes
 
+## Version 0.13.0
+### New Features
+- NuGet package now also deploys XML documentation file for RimuTec.Faker.
+
+
 ## Version 0.12.0
 ### New Features
 - Address.BuildingNumber()